pyproject.toml
[project]
name = "llm-code-review"
version = "0.1.0"
description = "Command line tool to code review GitHub pull requests with LLMs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "ollama>=0.4.7",
    "openai>=1.61.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
[project.urls]
Homepage = "https://github.com/geirfreysson/llm-code-review"
Changelog = "https://github.com/geirfreysson/llm-code-review/releases"
Issues = "https://github.com/geirfreysson/llm-code-review/issues"
CI = "https://github.com/geirfreysson/llm-code-review/actions"
[project.scripts]
llm-code-review = "llm_code_review.cli:cli"
